GeospatialCM/Geospatial/requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Fiona==1.7.12
Flask==1.1.1
future==0.18.2
gast==0.2.2
gcloud==0.17.0
GDAL==2.2.4
geocoder==1.38.1
geopandas==0.6.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.14.1
h5py==2.8.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jws==0.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mplleaflet==0.0.5
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.3
notebook==6.0.3
numpy==1.18.1
oauth2client==3.0.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyfcm==1.4.7
Pygments==2.5.2
pyparsing==2.4.6
pyproj==2.4.0
Pyrebase==3.0.27
pyrsistent==0.15.7
pytesseract==0.3.2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-jwt==2.0.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.6.0
ratelim==0.1.6
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
Rtree==0.9.3
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.6.4.post1
six==1.14.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.0.0
